New module spreadsheet_helper
